Multiple files under test in one test file
Put Unit into its own test file and rename Units.cpp to Quantity.cpp
